requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
groq
python-multipart
gTTS
